Refactor code: client now has the same processes.cpp file anf static funtions to create JsonObjects.
